pyproject.toml
[project]
name = "michizane-kun-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aenum==3.1.16",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "cryptography==45.0.5",
    "db-dtypes==1.4.3",
    "debugpy==1.8.15",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "executing==2.2.0",
    "flask==3.1.2",
    "frozenlist==1.7.0",
    "future==1.0.0",
    "google==3.0.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.176.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-bigquery==3.36.0",
    "google-cloud-bigquery-storage==2.33.0",
    "google-cloud-core==2.4.3",
    "google-crc32c==1.7.1",
    "google-genai==1.41.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.1",
    "grpcio-status==1.71.2",
    "gspread==6.2.1",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "line-bot-sdk==3.19.1",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "multidict==6.6.3",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandas-gbq==0.29.2",
    "pandas-stubs==2.3.2.250926",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydata-google-auth==1.9.1",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pywin32==311",
    "pyzmq==27.0.0",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "ruff>=0.14.11",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "types-pytz==2025.2.0.20250809",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "yarl==1.20.1",
]
